Adding more roles validations

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 use App\User;
 use App\Review;
 use Illuminate\Http\Request;
+use jeremykenedy\LaravelRoles\Models\Role;
+use jeremykenedy\LaravelRoles\Models\Permission;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
@@ -28,16 +30,26 @@ class UserController extends Controller
 
     public function delete(User $user)
     {
-    	$user->delete();
-
-    	return $user;
+        if($job->user_id == $request->user()->id || $request->user()->hasRole('admin')){
+        	$user->delete();
+        	return $user;
+        }else{
+            return response()->json([
+                'message' => 'You do not allowed to do this action'
+            ],403);
+        }
     }
 
     public function update(Request $request, User $user)
     {
-        $user->update($request->all());
-
-        return response()->json($user,200);
+        if($job->user_id == $request->user()->id || $request->user()->hasRole('admin')){
+            $user->update($request->all());      
+            return response()->json($user,200);
+        }else{
+            return response()->json([
+                'message' => 'You do not allowed to do this action'
+            ],403);
+        }
     }    
 
 }
